Move picard.ui.item to picard.item
- classes defined here are mainly used at top level
- they aren't specific to UI, as they are used to define important classes like Cluster, Album, etc...
- there's not even any Qt import in there
